+++ HISTORY.rst
@@ -2,6 +2,21 @@
 History
 =======
 
+2.3.0 (2018-09-26)
+
+New Features:
+* Added hexbin plot type.
+* More control over grouped axis label orientation.
+* Added alpha control to scatter, line, and parallel plots.
+* Added control over marker style to scatter plot.
+* Added ability to create custom color palettes.
+
+Changes:
+* Changed default accent color.
+* Visual tweaks to lollipop plot.
+* Bar plots with a few number of series will have better widths.
+
+
 2.2.0 (2018-09-17)
 ------------------